Added authentication logic

+++ controllers/contactsControllers.js
@@ -2,14 +2,18 @@ import HttpError from "../helpers/HttpError.js";
 import * as contactsService from "../services/contactsServices.js";
 
 export const getAllContacts = async (req, res) => {
-  const contacts = await contactsService.listContacts();
+  const { id: owner } = req.user;
+
+  const contacts = await contactsService.listContacts({ owner });
+
   res.json(contacts);
 };
 
 export const getOneContact = async (req, res) => {
+  const { id: owner } = req.user;
   const { id } = req.params;
 
-  const contact = await contactsService.getContactById(id);
+  const contact = await contactsService.getContact({ id, owner });
 
   if (!contact) {
     throw new HttpError(404);
@@ -19,8 +23,9 @@ export const getOneContact = async (req, res) => {
 
 export const deleteContact = async (req, res) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
 
-  const contact = await contactsService.removeContact(id);
+  const contact = await contactsService.removeContact({ id, owner });
 
   if (!contact) {
     throw new HttpError(404);
@@ -29,18 +34,20 @@ export const deleteContact = async (req, res) => {
 };
 
 export const createContact = async (req, res) => {
+  const { id: owner } = req.user;
   const data = req.body;
 
-  const contact = await contactsService.addContact(data);
+  const contact = await contactsService.addContact({ ...data, owner });
 
   res.status(201).json(contact);
 };
 
 export const updateContact = async (req, res) => {
+  const { id: owner } = req.user;
   const { id } = req.params;
   const data = req.body;
 
-  const contact = await contactsService.updateContact(id, data);
+  const contact = await contactsService.updateContact({ id, owner }, data);
 
   if (!contact) {
     throw new HttpError(404);
